Moved navigation elements to own `navigation` folder
