+++ frontend/src/components/navigation/Logo.vue
