+++ frontend/src/components/navigation/MobileSearch.vue
