+++ frontend/src/components/navigation/NavDesktopSearch.vue
