+++ frontend/src/components/navigation/NavDesktopUserMenu.vue
